libdf: Move dataloder stuff into new module
